Вносит правки в модуляции
